removed linux dependencies
svn path=/trunk/iCub/contrib/src/eMorph/; revision=18536

+++ src/basicCartesianCollector/src/plotterThread.cpp
@@ -325,8 +325,8 @@ int plotterThread::integrateImage(ImageOf<PixelMono>* imageIn, ImageOf<PixelMono
             pimageth += padding;
         }
         if((countX!=0) && (countY!=0)) {
-            centroidX = (int) floor(sumX / countX);
-            centroidY = (int )floor(sumY / countY);
+            centroidX = (int) floor((double) sumX / countX );
+            centroidY = (int) floor((double) sumY / countY );
         }
     }
     return  centroidY * retinalSize + centroidX;
